requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
Pillow==10.3.0
imagehash==4.3.1
streamlit>=1.36.0
plotly==5.22.0
pandas==2.2.2
pydantic==2.7.1
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.20.0
google-generativeai==0.5.4
motor==3.4.0
